requirements.txt
# Legion v4 — full-stack AI workstation control + multi-agent platform

# Telegram bot
aiogram>=3.4.0,<4.0.0
python-dotenv>=1.0.0

# LLM — all cloud providers + tool calling
# 1.57+ required for Cerebras Qwen-3-235b-a22b + Groq Kimi-K2 support
litellm>=1.57.0

# HTTP client (used by litellm + web tools)
httpx>=0.27.0

# Web scraping (legacy fallback)
requests>=2.31.0
beautifulsoup4>=4.12.0

# Vision / images
pillow>=10.0.0

# Async utilities
aiofiles>=23.2.0
aiohttp>=3.9.0

# System monitoring
psutil>=5.9.0

# ── Phase 1: Web browsing (JS-rendered) ──────────────────────────────────
playwright>=1.40.0

# ── Phase 2: Persistence ─────────────────────────────────────────────────
aiosqlite>=0.20.0

# ── Phase 3: Document processing ─────────────────────────────────────────
openpyxl>=3.1.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
python-docx>=1.0.0

# ── Phase 4: Email ───────────────────────────────────────────────────────
aioimaplib>=1.1.0
aiosmtplib>=3.0.0

# ── v4: RSS feeds ────────────────────────────────────────────────────────
feedparser>=6.0.0

# ── arXiv paper download ─────────────────────────────────────────────────
arxiv>=2.1.0

# ── WAJAR_WATCH: regulation monitoring pipeline ────────────────────────────
pdfminer.six>=20221105
PyGithub>=2.1.1
supabase>=2.3.0,<3.0.0
python-dateutil>=2.9.0

# ── Host machine tools (install separately, not pip) ──────────────────────
# sudo apt install xdotool wmctrl scrot xclip xdg-utils
# sudo apt install tesseract-ocr tesseract-ocr-eng tesseract-ocr-ind
# playwright install chromium
# pip install pip-audit  (optional, for /vuln_scan)
